tests: run the suite on hosts without netns, root, or Linux-only backends
- Convert the genuinely networked suites (REST, SMB with exec'd
  clients, S3, daemon, mount_create) to chimera_add_netns_test(): netns
  isolation where available, host network serialized under the
  chimera_net RESOURCE_LOCK where not.
- Gate linux-backend registrations on CHIMERA_LINUX across the posix
  and client matrices; previously several hundred tests registered for
  a module that only builds on Linux.
- Session roots resolve portably everywhere they were hard-coded:
  $CHIMERA_TEST_ROOT override, else the devcontainer's /build/test,
  else /tmp/chimera_test (posix_test_common.h, client_test_common.h,
  fsx.c, run_fsx.sh).
- Fixture chowns only attempt when actually root, and fixtures disable
  the S3 listener where a suite never speaks S3 (port 5000 belongs to
  ControlCenter on macOS).
- fsx compiles and runs off Linux: loff_t/fallocate ride chimera's
  POSIX shim, O_DIRECT and the fallocate range-mode probes are guarded,
  and test_dontcache_io() probes through chimera_posix_preadv2 (the
  raw preadv2 call was probing a host fd, not the chimera one).
- io_serialize gets its longer timeout on cairn as well as diskfs: it
  runs in the parallel pack now instead of holding the old exclusive
  net lock.
- Tests missing their external client (smbclient) exit 77 and register
  SKIP_RETURN_CODE 77 so they skip instead of fail.
- test_limits_wrapper.sh stays quiet on hosts without /proc: a failed
  redirect is reported by the shell before a same-command 2>/dev/null
  applies, so the suppression wraps the command.

+++ src/posix/tests/test_pio.c
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2025 Chimera-NAS Project Contributors
+// SPDX-FileCopyrightText: 2025-2026 Chimera-NAS Project Contributors
 //
 // SPDX-License-Identifier: LGPL-2.1-only
 
@@ -199,7 +199,7 @@ test_preadv_pwritev(int fd)
     off_t amt = chimera_posix_preadv(fd, iov, 2, 20);
 
     if (amt != 6) {
-        fprintf(stderr, "preadv failed: expected 6, got %zd\n", amt);
+        fprintf(stderr, "preadv failed: expected 6, got %lld\n", (long long) amt);
         exit(1);
     }
